Detach Settings into UserSettingsMixin, filter->filter_string
