+++ gvm/protocols/gmpv208/gmpv208.py
@@ -36,7 +36,6 @@ from gvm.protocols.base import GvmProtocol
 
 from gvm.utils import (
     check_command_status,
-    to_base64,
     to_bool,
     add_filter,
 )
@@ -237,41 +236,6 @@ class GmpV208Mixin(GvmProtocol):
 
         return self._send_xml_command(cmd)
 
-    def get_settings(self, *, filter: Optional[str] = None) -> Any:
-        """Request a list of user settings
-
-        Arguments:
-            filter: Filter term to use for the query
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_settings")
-
-        if filter:
-            cmd.set_attribute("filter", filter)
-
-        return self._send_xml_command(cmd)
-
-    def get_setting(self, setting_id: str) -> Any:
-        """Request a single setting
-
-        Arguments:
-            setting_id: UUID of an existing setting
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_settings")
-
-        if not setting_id:
-            raise RequiredArgument(
-                function=self.get_setting.__name__, argument='setting_id'
-            )
-
-        cmd.set_attribute("setting_id", setting_id)
-        return self._send_xml_command(cmd)
-
     def get_system_reports(
         self,
         *,
@@ -391,45 +355,6 @@ class GmpV208Mixin(GvmProtocol):
             _xmlauthconf = _xmlgroup.add_element("auth_conf_setting")
             _xmlauthconf.add_element("key", key)
             _xmlauthconf.add_element("value", value)
-
-        return self._send_xml_command(cmd)
-
-    def modify_setting(
-        self,
-        setting_id: Optional[str] = None,
-        name: Optional[str] = None,
-        value: Optional[str] = None,
-    ) -> Any:
-        """Modifies an existing setting.
-
-        Arguments:
-            setting_id: UUID of the setting to be changed.
-            name: The name of the setting. Either setting_id or name must be
-                passed.
-            value: The value of the setting.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not setting_id and not name:
-            raise RequiredArgument(
-                function=self.modify_setting.__name__,
-                argument='setting_id or name argument',
-            )
-
-        if value is None:
-            raise RequiredArgument(
-                function=self.modify_setting.__name__, argument='value argument'
-            )
-
-        cmd = XmlCommand("modify_setting")
-
-        if setting_id:
-            cmd.set_attribute("setting_id", setting_id)
-        else:
-            cmd.add_element("name", name)
-
-        cmd.add_element("value", to_base64(value))
 
         return self._send_xml_command(cmd)
