+++ gvm/protocols/gmpv214/__init__.py
@@ -117,6 +117,7 @@ from gvm.protocols.gmpv208.entities.tickets import (
     get_ticket_status_from_string,
 )
 from gvm.protocols.gmpv208.entities.tls_certificates import TLSCertificateMixin
+from gvm.protocols.gmpv208.entities.user_settings import UserSettingsMixin
 from gvm.protocols.gmpv208.entities.users import (
     UserAuthType,
     get_user_auth_type_from_string,
@@ -173,6 +174,7 @@ class Gmp(
     ScannersMixin,
     SchedulesMixin,
     SecInfoMixin,
+    UserSettingsMixin,
     UsersMixin,
 ):
     def __init__(
