+++ tests/protocols/gmpv208/testcmds/test_get_settings.py
@@ -23,8 +23,8 @@ class GmpGetSettingsTestCase:
 
         self.connection.send.has_been_called_with('<get_settings/>')
 
-    def test_get_settings_with_filter(self):
-        self.gmp.get_settings(filter="foo=bar")
+    def test_get_settings_with_filter_string(self):
+        self.gmp.get_settings(filter_string="foo=bar")
 
         self.connection.send.has_been_called_with(
             '<get_settings filter="foo=bar"/>'
